Removed unnecessary messy components
