+++ package.json
@@ -24,6 +24,7 @@
     "next": "13.1.6",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-router-dom": "^6.8.1",
     "typescript": "4.9.5"
   }
 }
