+++ src/pages/collection.tsx
@@ -2,11 +2,13 @@ import Head from 'next/head';
 import { Inter } from '@next/font/google';
 import Container from '@mui/material/Container';
 import Sidebar from '@/components/sidebar';
-import Collection from '@/components/collection';
+import { useUser } from '@auth0/nextjs-auth0/client';
+import { Background, StyledMainBorder } from '@/styles/Background';
 
 const inter = Inter({ subsets: ['latin'] })
 
-export default function Home() {
+const UserCollection = () => {
+  const { user, isLoading } = useUser();
   return (
     <>
       <Head>
@@ -16,9 +18,19 @@ export default function Home() {
       <main>
         <Container maxWidth="xl" sx={{ display: 'flex', p:0}} disableGutters>
           <Sidebar/>
-          <Collection/>
+          <Background aria-label='collection-background'>  
+            {user ? 
+              <StyledMainBorder aria-label="collection-box">
+              <h2>This is where the stuff will go to get and display the collection</h2>
+            </StyledMainBorder> 
+            // If no user then render an empty box
+            : <StyledMainBorder></StyledMainBorder>}
+
+          </Background>
         </Container>
       </main>
     </>
   )
 }
+
+export default UserCollection;
