--- src/utils/pageIndex.tsx
@@ -1,9 +1,0 @@
-const pageIndex = [
-  {name:'Account', id:'account'},
-  {name:'Collection', id:'collection'},
-  {name:'Community', id: 'community'},
-  {name:'High Scores', id: 'highscores'}
-]
-
-
-export default pageIndex;
